some refactoring on policy builders
