+++ libs/apollo/src/lib/cachePolicy/utils/index.ts
@@ -1,0 +1,2 @@
+export { buildPolicyConfig } from './buildPolicyConfig';
+export { queryPolicyConfig } from './queryPolicyConfig';
